requirements.txt
mysql-connector-python >= 8.0.28, <=8.0.29

coverage>=7.0,<8.0

django-mysql>=4.4.0,<4.5.0

Django>=3.2,<4.0

djangorestframework>=3.12.2,<3.13.0

djangorestframework-guardian2

django-model-utils>=4.3.1,<4.4.0

django-notifications-hq >= 1.8.3, < 2.0

django-filter>=23.5,<24.0

django-renderpdf >=4.0, <5.0

drf-spectacular ==0.25.1

flake8>=7.0.0,<7.1.0

gunicorn>=20.0.4,<20.1.0

isort>=5.13.2,<5.14.0

psycopg2-binary >= 2.9, < 3.0

weasyprint==52.5

docutils>=0.20.1